Added missing Bind*() from captureless lambdas and BindOnce from RepeatingCallback.
base::Bind*() functions were missing:
- binding OnceCallbacks from captureless lambdas
- binding RepeatingCallbacks from captureless lambdas
- binding OnceCallbacks from RepeatingCallbacks

This commit adds this functionality. It is implemented by adding another Bind*()
function with SFINAE enabled only for functors which are not pointers (to
disable it for function pointers) and where `base::Bind*(+functor, args...)`
was a valid expression (which is popular trick for explicit conversion from
captureless lambda to function pointer)

This commit also adds lots of new unit tests for new functionality and copies of
existing BindRepeating() tests but now for BindOnce() as well.