Integrate glog logging system into the project and use it in example.
This commits integrates glog third-party library as the logging system and
sets up everything to use it later on throughout the code.
Rework of how CMake targets were defined was needed to achieve this (as glog
doesn't build with the same warnings as intended for this library). The new
configuration should be easier to maintain in the future.