Implement support for binding to `base::WeakPtr` and executing weak callbacks.
This commit adds support for binding to `base::(Once|Repeating)Callback` objects
to weak pointers in the form of `base::WeakPtr<T>` objects. With this, users can
now create weak functions which will not be invoked if the weak pointers will be
invalidated. This is especially useful when posting tasks between threads and
sequences (though one must remember to ensure that bounded weak pointers will
be dereferenced (during callback execution) on the correct threads/sequences).