Fix compilation on MSVC compiler.
This was mostly due to complex SFINAE logic in base::Bind*()
functions when detecting lambdas. To make it easier on the compiler
the detection heuristic has been slightly simplified. It should work
for now, but in the future it might need some fine-tuning.

Other than that, it also required fixing a few signed-unsigned
comparisons and fixing GTest compilation options to ensure that
same crt options are used for project and GTest.