Added .clang-tidy (w/ CMake integration) and fixed build issues
Added .clang-tidy configuration file which lists many checks to be
performed. All failed checks result in compile errors. List of checks
is to be verified in the future and may grow or be reduced after
some consideration.

Clang-tidy is also integrated with CMake so it is enabled by default
when building with Clang toolchain.

Commit consists also of multiple smaller fixes throughout the
codebase required to keep building it (mostly missing `explicit`
keyword for constructors and conversion operators).