Adjusted Codecov settings to allow small drop in coverage
This commit changes Codecov settings to allow a small drop (%) in
coverage to still be counted as a success. This is to avoid
false negatives when well-tested code was refactored into smaller
code which could slightly reduce coverage percentage.

Current setting is chosen arbitraly to allow 2% drop in a single
commit (from the parent) and might be adjusted further in the future.