Support binding lambdas through FunctorTraits (vs conditions)
This commit changes the way lambdas are handled when passed to the
`base::Bind*()` functions from condition-based (via `if constexpr()`)
to FunctorTraits-based. This allows simpler implementation and easier
composition of new and different types of functors that may be bound.

It is also a first step/proof-of-concept for how [#11] can be
implemented without hardcoding all the helpers everywhere.