[#15] Add support for argument wrappers when binding callbacks
This commit adds support for wrapping and preserving wrappers when
binding arguments to base::*Callback objects. Currently only one
wrapper is supported: std::reference_wrapper<> which is now needed
to be used when binding something to non-const reference function
parameter. Binding raw values to const-reference parameters is still
allowed but a copy will be used (hence modifications to original
as well as it's lifetime value will affect the callback).

This commit also replaces non-const reference parameters in bounded
functions to pointers to avoid the check mentioned above.