[#16] Implement base::SplitOnceCallback helper
This commit adds implementation of the `baase::SplitOnceCallback()`
helper that can be used to split single OnceCallback into two
separate once callbacks. This may be useful when providing a callback
to API that requires separate success/failure callbacks or to pass
a callback into a function which might or might not run it.

NOTE:
  One difference is made from original //base version - this
  implementation allows splitting OnceCallbacks that do return
  something.