[#17] Add functionality to queue last task when stopping a thread
This commit adds an overload of Thread::Stop() which takes
a callback as a last task that is meant to be queued right before
stopping that thread. This can be used by users do perform
cleanup of thread-affine objects that are normally expected to
outlive the thread, but they are supposed to be destroyed on that
thread too. With this approach, users can be sure that nothing
else will use these objects (on that thread) so in some cases
it may be a valid approach to bind callbacks to Unretained()
wrapper.