[#7][#19] Implement base::TimeTicks and ass missing related helpers
This commit adds `base::TimeTicks` implementation of steady clock
that can be used to measure time spent on something without the
worry that an external clock can be modified in the meantime.

Furthermore, it adds some of the missing helpers (methods and
operators) for Time and TimeDelta classes (while also providing
them for new TimeTicks class).