[#24] Add single libbase init/deinit logic and customize log format
This commit adds single logic for initialization and
deinitialization of the libbase library:

* base::Initialize(...),
* base::Deinitialize().

These should be called prior to any possible usage of the library
and after the last call to any libbase object, respectively.

Furthermore, this commit adds custom log formatter for GLOG to
match default's Chromium log format (without PID, which would
be unused for now).