[#13] Add unit tests for base::DelayedTaskManager
This commit adds unit tests for the `base::DelayedTaskManager` and
`base::DelayedTaskManagerSharedInstance` classes. Along with tests,
a simple fixes or small improvements are also made to the logic to
improve performance and latency of scheduling delayed tasks.

There is also a change in how DelayedTaskManager objects are
retrieving current time (TimeTicks) - which is now injected via
the constructor so that the unit tests can mock time progression.