Ftr/trace events/1 (#29)
* [#9] Implement simple tracing framework (trace events)

This commit introduces simple tracing framework based on what
Chromium offers. Similar macros are exposed to the users but with
variadic arguments (no need to change macro name to add/remove
arguments). For now, the only configuration present is either
to enable or disable all traces. Furthermore, current implementation
is quite basic, meaning that there is still a lot of room for
performance and memory-related optimizations down the road.

Currently these tracing options are available:
* Trace event (begin+end or per-scope)
* Async trace events (begin/step/end)
* Trace counters
* Trace events with flow ("arrows")
* Trace signals