[#9] Make tracing opt-in instead of opt-out (default opt-in)
This commit reverses the configuration for enable-if-not-disabled
to enable-if-enabled (and keeps it enabled by default). This should
make it easier to follow the code and will be used for other
optional features in the future.

Configuration is done on the CMake level. Builing `libbase` with
tracing disabled can be done with:

  cmake -DLIBBASE_FEATURE_TRACING=OFF -S . -B build
  cmake --build build/

Also added few missing #ifdefs to ensure that everything builds
correctly with tracing disabled.