[#27] Setup documentation generation system
This commit adds basic sysytem for library documentation generation.
The final documentation will be generated with Sphinx to achieve
nice-looking and easy-to-use documentation. The source data is
collected by its extension Breath that takes provided XML data
and converts it to RST which in turn is converted to nice HTML.
The XML input data is collected from Doxygen which is automatically
invoked by another Sphinx extension called Exhale.

The documentation can be built manually with:

  cd docs
  sphinx-build -b html . build

or with CMake:

  cmake -DLIBBASE_BUILD_DOCS=ON -S . -B build
  cmake --build build --target docs

The generated documentation will be available in:

  libbase/docs/build/

directory.

Currently, the documentation consists of only automatically
generated content (based on Doxygen's output) but in the future
it will contain manually written documentation as well.

Furthermore, this adds CI job to build documentation for known
branches and tags and publish it to GitHub Pages.