Print output of failed test runs of unit tests on CI
This commit enables CI runners to print the output of the failed
test runs. This is to improve chances of reproducing and fixing
the bugs in the future, as it may contain - for example - stack
traces when the test crashed.

This was triggered by the [#30] bug report which was a one-shot
crash that couldn't be reproduced locally and there were no
information about where it crashed at all.

[no ci]