[#27] Added "Logging" documentation that points to GLOG's user guide
This commit adds "Logging" documentation page. This page now points
to GLOG's user guide (on GitHub) where the up-to-date version is
presented.
For ease-of-use and to ensure that `libbase` users will have correct
version available, the GLOG's `README.rst` file from the

  third_party/glog/

directory is directly included into the documentation to present
user guide from the same version of the GLOG library that is being
used by the library itself (e.g. for older builds).