[CI] Checkout submodules too when building documentation
This commit changes documentation build&deploy workflow to checkout
submodules along with the main repository. This is needed to be
able to include some parts of their documentation/readmes files/...
inside the documentation.

Currently, "Logging" page includes parts of the "User guide" of the
GLOG library (from checkouted submodule in third_party/glog/).