Drop gcc on MacOS CI builds as its not reliable, test newer versions (#43)
This commit drops MacOS builds with GCC in CI as support for GCC is not
reliable enough. For simplicity let's stick with just default clang
version.

Also include builds on MacOS 14 and latest - intended to be 15 but at
the moment its in preview and not fully reliable (e.g. missing things like `ninja`).