Implement base::BindToCurrentSequence helper (#44)
This commit adds a new helper that binds a callback execution to
sequence that code binding it is executed on. This means that when the
resulting callback will be executed, it will post-task execution of the
original one to the task runner/sequence on which it wound bound (to
"current" - at the time - sequence).

This is particularly useful if you're passing a result/on-done callback
to another object which could work on different thread/sequence and you
want to ensure that it is executed on your original sequence.