[#41] Implement network request cancellation
This commit adds a functionality to abort/cancel in-flight network
request (on best efforts basis - asynchronously). When a network request
is created, a cancellation token will be return which can be used to try
to cancel the request.

Due to nature of the implementation (asynchronous), the operations is
working on best-effort basis - it may be possible that request's on-done
callback will still be invoked after cancellation of the request.
To ensure it doesn't happen, bind callback to a weak pointer that you
can invalidate when needed. As such, this is should be considered to
mostly being used to avoid wasting resources on a request for which a
response is no longer needed.