[#41] Provide task runner to bind response callback or auto-bind to current
This commit adds a functionality to SimpleUrlLoader that allows user to
specify on which task runner should the response for the given request
(via `on_done_callback`) should be executed. If no task runner is
provided it will bind it to current sequence (which is usually what
users would like).

This commit also:

* added Timeout error code
* simplified networking examples with above + using built-in timeout
  handling instead of using guard delayed tasks
* fixed a bug where users couldn't use header-only mode with POST
  requests