Update README and documentation with new features and changes
This commit updates the README and project documentation to reflect
new features introduced in the latest commits as well as some changes.

It also updates the information about building and installing the
project with `vcpkg` and fixes the `vcpkg` port name to be used:

  `ripper37-libbase`

Additionally it adds a vcpkg shield/badge in README to advertise library
availability in that package manager.