[#41] Make net module optional and separate target (`libbase::libbase_net`)
This commit changes the structure of the `net` module so that it can be
made optional. New build option is introduced to allow specifying
whether to build and use the new `net` module or not:

  LIBBASE_BUILD_MODULE_NET

Additionaly, minor changes to initialization logic were made - now users
have to initialize separate modules independently.
Furthermore the simple example has been updated to work correctly
regardless of whether the `net` module is built or not by using
conditional linking and newly introduced compiler definition for library
users: `LIBBASE_MODULE_NET`.

This will be made as separate feature (probably enabled by default) for
vcpkg users, that can be disabled to avoid pulling in `libcurl`.