[wx] Add wxWidgets integration with example application
This commit adds implementation of utilities to integrate `libbase` with
wxWidgets based applications. Main component available for users is the
new class `wxMessageLoopAttachment`, which bridges the wxWidgets event
loop with the `libbase` threading/task runner/message loop systems which
in turn allows you to - among others - fetching "current's thread's"
task runner (even if this is main/wxWidgets thread), posting tasks to it
(these will be executed as if you'd post events to specified
`wxEventHandler` object) and so on.

Recommended usage is to create an instance of `wxMessageLoopAttachment`
when initializing application/wxWidgets thread and destroy it on
shutdown. See `wxwidgets_integration` example for more details.