[#41] Implement base::net::UrlRequest for advanced use cases (streaming)
This commit adds a new network class `base::net::UrlRequest` that can be
used in more advanced scenarios. Contrary to simple url loader usage,
here the user receives multiple callbacks on each phase of the request:

- when response starts (with metadata such as result code and headers)
- when new data chunk is fetched (for every downloaded chunk)
- when request finishes (either succesfully or not)

Only the last call is guaranteed to be performed, so users should not
depend on receiving `OnResponseStarted` or `OnWriteData` calls in their
applications.

Users are required to buffer the data by themselves and cancel the
requests when needed. Once `UrlRequest::Cancel` method is invoked (or
the `UrlRequest` object is destroyed), it is guaranteed that no more
calls to passed `UrlRequest::Client` will be made (not even
`OnRequestFinished`) and the request will be asynchronously cancelled at
first chance.

Additionally, new example called `networking` was added and earlier
additions to `simple` example related to networking module were moved to
it along with new example that showcases usage of `UrlRequest` class.