Update vcpkg config, CMake script and documentation
This commit overhauls CMake script to improve building and installation
of the library. It also updates the vcpkg configuration.

Lastly, the documentation and README files are now updated to better
describe how to build the library and how to consume it in other
projects.

One of the notable CMake script improvements were on building and
consuming the library. You can now ask `libbase` to automatically set up
its own `vcpkg` installation to resolve its dependencies during the
build. While it's not the recommended way to manage your dependencies,
the library is also improved compatibility for consuming it with CMake's
`FetchContent` and `add_subdirectory` commands. Note that these do NOT
auto-resolve dependencies so you will have to add them as well.
They are not handled by `libbase` itself in this case to ensure that
the same version will be provided for `libbase` and the project using
it.

----

One potentially breaking change in how to consume the library is that
for optional modules (net/win/wx/...) you will now have to explicitly
ask CMake to find targets for it to make those extra targets available:

```cmake
find_package(libbase CONFIG REQUIRED [COMPONENTS] <optional_components>...)
```

where `<optional_components>` is e.g. `net wx`. Targets are always named
in `libbase::libbase_${component_name}` format. Be sure to build and
install the component/module before using it.