Add module ifdef guards and helper methods to net::ResourceRequest
This commit adds #ifdef guards to all headers that are for optional
modules to ensure that code in them is not processed if the module is
not built/installed. At the moment there are no errors, but this may
change in the future.

Furthermore, this commit also adds some convenience helper methods to
`base::net::ResourceRequest` to make it easier to create them.
Previously user had to provide all parameters until the last intended
one (due no designed initializers in C++17). With new methods, users
can start with URL and then add parameters as they see fit. Example:

```cpp
    base::net::ResourceRequest{"https://www.google.com/robots.txt"}
        .WithHeadersOnly()
        .WithTimeout(base::Seconds(5)),
```