Fix compilation on MSVC 2019 (#60)
This commit fixes the compilation on older MSVC 2019 (19.29) by converting two
warnings that were treated as errors back to warnings. This problems doesn't occur
on newer MSVC 2022 and probably is resolved on glog's `master` or in one of the
maintained forks (`ng-log` or Abseil Logging).

For `v1.2.0` we should probably update/change the dependency to one of those
to stay with maintained dependencies.