Fix bug causing increased CPU usage when base::Thread is idle
Recently (in #49) a bug was introduced that changed behavior of
`base::Thread`/`base::MessageLoop` which is idle (no tasks to execute).
Previously it was waiting for a next task and after the change instead
of waiting (asleep) it constantly checked if a new task is ready (busy
loop) which caused increased CPU usage on that thread.

This commit fixes it by ensuring that MessageLoop::Run() will always
wait for the next task.